requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
Brotli==1.1.0
browser-cookie3==0.20.1
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clr_loader==0.2.7.post0
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
cssselect2==0.8.0
curl_cffi==0.11.4
defusedxml==0.7.1
Deprecated==1.2.18
duckduckgo_search==8.0.4
et_xmlfile==2.0.0
fastapi==0.115.14
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
g4f==0.5.6.3
h11==0.16.0
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
lz4==4.4.4
magika==0.6.2
mammoth==1.9.1
markdownify==1.1.0
markitdown==0.1.2
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
mss==10.0.0
multidict==6.6.3
nest-asyncio==1.6.0
nodriver==0.46.1
numpy==2.3.1
olefile==0.47
onnxruntime==1.22.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfminer.six==20250506
pillow==11.2.1
platformdirs==4.3.8
plyer==2.1.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
python-socks==2.7.1
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32==310
requests==2.32.4
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
starlette==0.46.2
sympy==1.14.0
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
WMI==1.5.1
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
